For that, instead of taking key prefix length and byte array, take the length
and source node to take the prefix bytes from, and copy them directly in words.

At the same time fix key_prefix_capacity to 7 as the current implementation is
suited for that length only anyway.

Performance change, micro_benchmark_key_prefix:

Baseline:

2020-06-24 20:58:16
Running ./micro_benchmark_key_prefix
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.54, 0.43, 0.19
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
unpredictable_get_shared_length          0.830 us        0.834 us       839479 items_per_second=25.1903M/s
unpredictable_leaf_key_prefix_split       21.4 us         21.4 us        32690 items_per_second=11.8017M/s
unpredictable_cut_key_prefix              24.9 us         24.8 us        28262 items_per_second=10.1472M/s

With the change:

unpredictable_get_shared_length          0.826 us        0.826 us       847066 items_per_second=25.4101M/s
unpredictable_leaf_key_prefix_split       21.2 us         21.2 us        33093 items_per_second=11.9012M/s
unpredictable_cut_key_prefix              21.3 us         21.3 us        32893 items_per_second=11.8406M/s

U Test:

Benchmark                                                    Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------------------------------
unpredictable_get_shared_length                           -0.0125         -0.0149             1             1             1             1
unpredictable_get_shared_length                           -0.0127         -0.0151             1             1             1             1
unpredictable_get_shared_length                           -0.0126         -0.0148             1             1             1             1
unpredictable_get_shared_length                           -0.0127         -0.0147             1             1             1             1
unpredictable_get_shared_length                           -0.0125         -0.0144             1             1             1             1
unpredictable_get_shared_length                           -0.0128         -0.0147             1             1             1             1
unpredictable_get_shared_length                           -0.0127         -0.0144             1             1             1             1
unpredictable_get_shared_length                           -0.0127         -0.0143             1             1             1             1
unpredictable_get_shared_length                           -0.0126         -0.0144             1             1             1             1
unpredictable_get_shared_length_pvalue                     0.0004          0.0004      U Test, Repetitions: 9 vs 9
unpredictable_get_shared_length_mean                      -0.0126         -0.0146             1             1             1             1
unpredictable_get_shared_length_median                    -0.0127         -0.0147             1             1             1             1
unpredictable_get_shared_length_stddev                    +0.3368         +2.6749             0             0             0             0
unpredictable_leaf_key_prefix_split                       -0.0107         -0.0111            21            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0154         -0.0157            22            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0151         -0.0156            22            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0136         -0.0137            22            21            22            21
unpredictable_leaf_key_prefix_split                       -0.0145         -0.0147            22            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0140         -0.0142            21            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0138         -0.0140            21            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0090         -0.0096            21            21            21            21
unpredictable_leaf_key_prefix_split                       -0.0141         -0.0143            21            21            21            21
unpredictable_leaf_key_prefix_split_pvalue                 0.0004          0.0004      U Test, Repetitions: 9 vs 9
unpredictable_leaf_key_prefix_split_mean                  -0.0134         -0.0136            21            21            21            21
unpredictable_leaf_key_prefix_split_median                -0.0133         -0.0136            21            21            21            21
unpredictable_leaf_key_prefix_split_stddev                -0.3599         -0.3234             0             0             0             0
unpredictable_cut_key_prefix                              -0.0767         -0.0768            25            23            25            23
unpredictable_cut_key_prefix                              -0.0789         -0.0790            25            23            25            23
unpredictable_cut_key_prefix                              -0.0772         -0.0773            25            23            25            23
unpredictable_cut_key_prefix                              -0.0771         -0.0769            25            23            25            23
unpredictable_cut_key_prefix                              -0.0777         -0.0778            25            23            25            23
unpredictable_cut_key_prefix                              -0.0783         -0.0781            25            23            25            23
unpredictable_cut_key_prefix                              -0.0781         -0.0779            25            23            25            23
unpredictable_cut_key_prefix                              -0.0796         -0.0795            25            23            25            23
unpredictable_cut_key_prefix                              -0.0814         -0.0812            25            23            25            23
unpredictable_cut_key_prefix_pvalue                        0.0004          0.0004      U Test, Repetitions: 9 vs 9
unpredictable_cut_key_prefix_mean                         -0.0783         -0.0783            25            23            25            23
unpredictable_cut_key_prefix_median                       -0.0789         -0.0790            25            23            25            23
unpredictable_cut_key_prefix_stddev                       -0.5946         -0.5605             0             0             0             0